requirements.txt
streamlit>=1.39,<2
openai>=1.12.0
python-dotenv>=1.0.0
